Added functionality for improved search using latitude and longitude parameters, a session module for managing user sessions, and a new helper function for validating latitude and longitude inputs.

+++ app/car/factory/create_factory.go
@@ -33,9 +33,9 @@ func (f *CarFactory) CreateSqlFromGrpc(req *obdv1.CarCreateRequest) (*db.CarCrea
 func (f *CarFactory) CreateGrpcFromSql(req *db.Car) (*obdv1.CarCreateResponse, error) {
 	response := &obdv1.CarCreateResponse{
 		Car: &obdv1.Car{
-			CarId:           req.CarID,
-			CarBrandModelId: req.CarBrandModelID,
-			UserId:          req.UserID,
+			CarId:           int32(req.CarID),
+			CarBrandModelId: int32(req.CarBrandModelID),
+			UserId:          int32(req.UserID),
 			ModelYear:       req.ModelYear,
 			CreatedAt:       timestamppb.New(req.CreatedAt),
 		},
